cache and Moved files to main folder
